Refactor + SwiftUI Support
