+++ Example/NumericLabelExample/NumericLabelApp.swift
@@ -1,0 +1,17 @@
+//
+//  NumericLabelApp.swift
+//  NumericLabel
+//
+//  Created by 秋星桥 on 2025/1/6.
+//
+
+import SwiftUI
+
+@main
+struct NumericLabelApp: App {
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+        }
+    }
+}
