+++ Sources/NumericTransitionLabel/NumericTransitionLabel/Filter/CAFilter.swift
@@ -1,7 +1,7 @@
 //
 //  Created by ktiays on 2024/11/17.
 //  Copyright (c) 2024 ktiays. All rights reserved.
-// 
+//
 
 import Foundation
