+++ Sources/NumericTransitionLabel/NumericTransitionLabel/Filter/GaussianBlurFilter.swift
@@ -7,9 +7,8 @@ import Foundation
 import QuartzCore
 
 struct GaussianBlurFilter {
-    
     static let inputRadiusKeyPath = "filters.gaussianBlur.inputRadius"
-    
+
     let effect: NSObject
 
     var inputRadius: Double {
