+++ Sources/NumericTransitionLabel/NumericTransitionLabel/NumericTransitionTextLayer+State.swift
@@ -1,0 +1,132 @@
+//
+//  NumericTransitionTextLayer+State.swift
+//  NumericTransitionLabel
+//
+//  Created by 秋星桥 on 2025/1/6.
+//
+
+import Foundation
+
+extension NumericTransitionTextLayer {
+    class State {
+        var frame: CGRect = .zero {
+            didSet {
+                if frame == presentationFrame {
+                    return
+                }
+                var animation: AnimationState<CGRect> = frameAnimation ?? .init(value: presentationFrame, velocity: .zero, target: frame)
+                animation.target = frame
+                frameAnimation = animation
+            }
+        }
+
+        var presentationFrame: CGRect = .zero {
+            didSet {
+                delegate?.updateFrame(with: self)
+            }
+        }
+
+        var frameAnimation: AnimationState<CGRect>?
+
+        var scale: CGFloat = 1
+        var scaleAnimation: AnimationState<CGFloat>?
+
+        var offset: CGFloat = 0
+        var offsetAnimation: AnimationState<CGFloat>?
+
+        var opacity: Float = 1 {
+            didSet { layer.opacity = opacity }
+        }
+
+        var opacityAnimation: AnimationState<Float>?
+
+        var blurRadius: CGFloat = 0 {
+            didSet {
+                layer.setValue(blurRadius, forKeyPath: GaussianBlurFilter.inputRadiusKeyPath)
+            }
+        }
+
+        var blurRadiusAnimation: AnimationState<CGFloat>?
+
+        var delay: TimeInterval = 0
+        var invalid: Bool = false
+        var isDirty: Bool = true
+
+        var isAnimating: Bool {
+            frameAnimation != nil
+                || scaleAnimation != nil
+                || offsetAnimation != nil
+                || opacityAnimation != nil
+                || blurRadiusAnimation != nil
+        }
+
+        var isVisible: Bool {
+            if let opacityAnimation {
+                return opacityAnimation.value >= 0.01
+            }
+            return opacity >= 0.01
+        }
+
+        let layer: CALayer
+
+        var range: NSRange = .init()
+        var character: Character?
+        var textBounds: CGRect?
+
+        weak var delegate: (any Delegate)?
+
+        init(layer: CALayer) {
+            self.layer = layer
+        }
+
+        func updateTransform() {
+            let transform = CATransform3DConcat(
+                CATransform3DMakeScale(scale, scale, 1),
+                CATransform3DMakeTranslation(0, offset * frame.height / 3, 0)
+            )
+            layer.transform = transform
+        }
+
+        enum AnimationType {
+            case appear
+            case disappear
+        }
+
+        private static let smallestScale: CGFloat = 0.4
+        private static let appearBlurRadius: CGFloat = 9
+        private static let disappearBlurRadius: CGFloat = 6
+
+        func configureAnimation(with type: AnimationType, countsDown: Bool = false) {
+            switch type {
+            case .appear:
+                scaleAnimation = .init(value: Self.smallestScale, velocity: .zero, target: 1)
+                scale = Self.smallestScale
+                let offset: CGFloat = countsDown ? -1 : 1
+                offsetAnimation = .init(value: offset, velocity: .zero, target: 0)
+                self.offset = offset
+                opacityAnimation = .init(value: 0, velocity: .zero, target: 1)
+                opacity = 0
+                blurRadiusAnimation = .init(value: Self.appearBlurRadius, velocity: 0, target: 0)
+                blurRadius = Self.appearBlurRadius
+                updateTransform()
+            case .disappear:
+                var scaleAnimation = scaleAnimation ?? .init(value: scale, velocity: .zero, target: Self.smallestScale)
+                scaleAnimation.target = Self.smallestScale
+                self.scaleAnimation = scaleAnimation
+
+                let offset: CGFloat = countsDown ? 1 : -1
+                var offsetAnimation = offsetAnimation ?? .init(value: self.offset, velocity: .zero, target: offset)
+                offsetAnimation.target = offset
+                self.offsetAnimation = offsetAnimation
+
+                var opacityAnimation = opacityAnimation ?? .init(value: opacity, velocity: .zero, target: 0)
+                opacityAnimation.target = 0
+                self.opacityAnimation = opacityAnimation
+
+                var blurRadiusAnimation = blurRadiusAnimation ?? .init(value: blurRadius, velocity: 0, target: Self.disappearBlurRadius)
+                blurRadiusAnimation.target = Self.disappearBlurRadius
+                self.blurRadiusAnimation = blurRadiusAnimation
+            }
+        }
+    }
+}
