+++ Sources/NumericTransitionLabel/Supplement/ArrayContainer.swift
@@ -1,14 +1,13 @@
 //
 //  Created by ktiays on 2024/11/17.
 //  Copyright (c) 2024 ktiays. All rights reserved.
-// 
+//
 
 import Foundation
 
 final class ArrayContainer<T>: Sequence, ExpressibleByArrayLiteral {
-
     private var array: [T]
-    
+
     var isEmpty: Bool {
         array.isEmpty
     }
@@ -40,12 +39,12 @@ final class ArrayContainer<T>: Sequence, ExpressibleByArrayLiteral {
     func first(where predicate: (T) throws -> Bool) rethrows -> T? {
         try array.first(where: predicate)
     }
-    
+
     @inlinable
     func removeAll(where shouldBeRemoved: (T) throws -> Bool) rethrows {
         try array.removeAll(where: shouldBeRemoved)
     }
-    
+
     func makeIterator() -> IndexingIterator<[T]> {
         array.makeIterator()
     }
